Some minor work

+++ core/views.py
@@ -1,13 +1,68 @@
 from django.shortcuts import render, redirect
+from django.http import HttpResponse
+from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
+from django.contrib.auth.models import User
+from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.forms import UserCreationForm
 from django.http import HttpResponse
 from rest_framework.decorators import api_view
 from rest_framework.response import Response
-from .models import Room, Topic
+from .models import Room, Topic, Message
 from .forms import RoomForm
 
 
 # Create your views here.
+def loginPage(request):
+    page = 'login'
+
+    if request.user.is_authenticated:
+        return redirect('home')
+    
+    if request.method == 'POST':
+        username = request.POST.get('username').lower()
+        password = request.POST.get('password')
+
+        try:
+            user = User.objects.get(username=username)
+        except:
+            messages.error(request, 'User does not exist!')
+
+        user = authenticate(request, username=username, password=password)
+
+        if user is not None:
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, 'Username Or Password does not exist!')
+
+    context = {'page': page}
+    return render(request, 'core/login_register.html', context)
+
+
+def logoutUser(request):
+    logout(request)
+    return redirect('home')
+
+
+def registerUser(request):
+    form = UserCreationForm()
+
+    if request.method == 'POST':
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit=False)
+            user.username = user.username.lower()
+            user.save()
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, 'An error has occurred during registration!')
+
+    return render(request, 'core/login_register.html', {'form': form})
+
+
 def home(request):
     q = request.GET.get('q') if request.GET.get('q') != None else ''
 
@@ -25,9 +80,21 @@ def home(request):
 
 def room(request, pk):
     room = Room.objects.get(id=pk)
-    context = {'room': room}
+    room_messages = room.message_set.all().order_by('-created')
+
+    if request.method == 'POST':
+        message = Message.objects.create(
+            user=request.user,
+            room=room,
+            body=request.POST.get('body')
+        )
+        return redirect('room', pk=room.id)
+    
+    context = {'room': room, 'room_messages': room_messages}
     return render(request, 'core/room.html', context)
 
+
+@login_required(Login_url='login')
 def createRoom(request):
     form = RoomForm()
 
@@ -40,9 +107,14 @@ def createRoom(request):
     context = {'form': form}
     return render(request, 'core/room_form.html', context)
 
+
+@login_required(Login_url='login')
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse('You are not allowed here!')
 
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
@@ -53,8 +125,14 @@ def updateRoom(request, pk):
     context = {'form': form}
     return render(request, 'core/room_form.html', context)
 
+
+@login_required(Login_url='login')
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse('You are not allowed here!')
+    
     if request.method == 'POST':
         room.delete()
         return redirect('home')
